Correct msgids in the homepage Jinja template
The mdgid passed to the gettext '_' function in the templated differed
from that in the POT file by some leading whitespace. For some reason,
`pybabel extract` had not caught this.

The template was hand-edited to reflect the indentation in the PO(T)
files. It may be deemed appropriate n future commits to revert the msgid
in the template and change the ones in the POT files instead for the
sake of style.