Use lazy_gettext, not gettext, outside requests
https://codein.withgoogle.com/dashboard/task-instances/6001407916244992/?sp-page=1
http://tickets.musicbrainz.org/browse/CB-188
http://tickets.musicbrainz.org/browse/CB-197

Previously, there were uses of `gettext` outside of requests. Strings
were evaluated immediately, outside the context of a request and thus a
language to use a translation of.

Uses of `gettext` were thus modifided to `lazy_gettext`, which delays
evaluation until the string is used, inside requests. This fixes the
relevant strings in the tickets linked to above.