Make column ordering of dump consistent/compatible
http://tickets.musicbrainz.org/browse/CB-160

Previous commits to disclude draft reviews and revisions from the public
dump (85c7d1cab2 and 5622bcd1a) changed the column ordering of the
"COMBINED" dump. To ensure compatibility with even older code, this commit
reverts the column ordering.

Before the above-referenced commits, the column orderings for the
"COMBINED" and "SEPARATE" dumps used to be inconsistent, one being
tab-separated and the other comma-separated and parenthesis-encapsulated.
This commit makes the "SEPARATE" dump follow the format of the "COMBINED"
dump. This does not break compatibility with, say, `manage.py export
importer` because typically the "COMBINED" dump file would be used to
import.

All of this was achieved by writing raw SQL queries instead of SQLAlchemy
ORM expressions. While this is undesirable, the ORM did not seem to
provide an easy way to configure column orderings without hacky text
replacement, doing which is not better than just using raw SQL queries.