Remove note about drafts' not being private
The note about drafts' not being private would be inapplicable if
previous commits to disclude drafts from the public dump are accepted.
Hence, the note is being removed.

Removing the note eliminates the need for the relevant internationalized
string. This commit does not regenerate the POT file because all string
extractions are to be done on the master branch.