Translate review language to current user locale
https://github.com/metabrainz/critiquebrainz/pull/23#discussion-diff-48316217

As opposed to the website language selector, where the language names
are for those who know the language, the displaying of reviews'
languages is for users who might not know the review's language. Thus,
it makes more sense to translate the language name into the current
user's locale than the review language's locale itself.