Replace navbar "About" link with language selector
http://tickets.musicbrainz.org/browse/CB-111

Previously, the language selector was located at the bottom of the page,
bad for visibility. Now, the “About” link in the top has been replaced
with it.

There was always a link to the about page in the footer, so the loss of
the about page in the navbar could be justified.