Add language selector back to the footer
Previously, the language selector had been moved from the footer to the
top navbar. This means that when the top navbar is collapsed, in narrow
devices, the language selector requires a click to be visible.

To remedy this, the language selector has been added back to the footer,
but it will only show up if the navbar language selector is hidden.