Fix bug in initializing CB test database
The flask app created before creating the database tables checks
whether the database is being created for tests and uses the
config file accordingly. However, after that we again create
the flask app but do not check whether test configuration should
be used. Therefore, the SQLALCHEMY_DATABASE_URI gets set
incorrectly. This is evident from failure to insert license in
the license table just after. I do not see why we need to create
the app twice, hence remove the second creation and fix the bug.