Fix usage of OPTIONS method
OPTIONS method is needed for CORS prefetch requests. @crossdomain
decorator needs OPTIONS in methods otherwise OPTIONS requests to the
route will return 405 NOT ALLOWED.

However, we don't need to add OPTIONS to each `methods` list of a
particular route i.e. if there is a function to handle GET and another
to handle POST for a route, just adding to OPTIONS to the GET function
is enough.