Add test config back
Removed in 13fc976b1e9 because we thought that the configuration was the
same in both development and test due to the docker isolation, however
the name of the musicbrainz database in test is different.
We could have manually modified the config in each test base class, but
it seems easier to just revert the test-specific config file.