Return a 404 response if the given entity isn't in MusicBrainz
CB-401
We originally had the ability to show a page about an "Unknown entity"
if an MBID didn't exist in the database. This was useful when CB
connected to an MB replica that may have been out of date by up to an
hour. This would also allow us to show a review if an entity had since
been deleted.
Now CB connects directly to the main MB site, so this isn't an issue.
For entities that have been deleted we decided to handle these on a
case-by-case basis: CB-408

This commit also introduces the use of the brainzutils-mb-sample-database
docker image for the musicbrainz data source. This is a real database
with real data which allows us to test real/missing MBIDs without
needing to mock methods or insert test data at the beginning of each
test. It also makes test starup a bit faster.