Check redirect asserts against a relative Location instead of absolute
The new HTTP RFC 9110 (section 10.2.2) shows that Location: headers
can be to a relative location and don't require a protocol and host.
This changed from RFC 2616 (section 14.30), and so werkzeug changed
this in https://github.com/pallets/werkzeug/issues/2352
Flask-testing hasn't been updated to fix the assertRedirects method,
so just check the redirect location manually.