Set up babel with a dummy domain in the API app
We share some code and variables between the website and webservice
which is translated with lazy_gettext.
If the babel extension is not installed in the flask app, the
evaluation of this variable causes an unhandled exception in babel.
In order to prevent this exception, configure babel with a domain which
is different to the default `messages` domain used in the website.
This will result in no translations being found, and the strings being
passed through to the WS in english.