Improve uWSGI and consul-template interactions on reloading
We fixed the `bind(): Address already in use [core/socket.c line 769]` error in LB
using `exit-on-reload = true` in uWSGI configuration. However, this prevents graceful
reloading.

I did some further investigation as noted in metabrainz/artwork-redirect#46 and have
a better fix in mind now.

For the command field in exec block of consul template configuration, use array format.
This executes the command directly without spawning a shell wrapper or setting a process
group id. Both of which will interfere with forwarding signals to the uWSGI process.